+++ targets/ublksrv_tgt.cpp
@@ -6,19 +6,16 @@
 #define ERROR_EVTFD_DEVID   0xfffffffffffffffe
 
 static void ublksrv_ctrl_data_init(struct ublksrv_ctrl_dev *cdev,
-		struct ublksrv_ctrl_data *data, bool recover)
+		bool recover)
 {
-	ublksrv_tgt_jbuf_init(cdev, &data->jbuf, recover);
-	sem_init(&data->queue_sem, 0, 0);
-	data->recover = recover;
-	ublksrv_ctrl_set_priv_data(cdev, data);
+	ublksrv_tgt_jbuf_init(cdev, &cdev->data->jbuf, recover);
+	sem_init(&cdev->data->queue_sem, 0, 0);
+	cdev->data->recover = recover;
 }
 
-static void ublksrv_ctrl_data_exit(struct ublksrv_ctrl_dev *cdev,
-		struct ublksrv_ctrl_data *data)
+static void ublksrv_ctrl_data_exit(struct ublksrv_ctrl_dev *cdev)
 {
-	ublksrv_ctrl_set_priv_data(cdev, NULL);
-	ublksrv_tgt_jbuf_exit(&data->jbuf);
+	ublksrv_tgt_jbuf_exit(&cdev->data->jbuf);
 }
 
 struct ublksrv_tgt_jbuf *ublksrv_tgt_get_jbuf(const struct ublksrv_ctrl_dev *cdev)
@@ -321,7 +318,6 @@ static int ublksrv_tgt_start_dev(struct ublksrv_ctrl_dev *cdev,
 
 static int ublksrv_device_handler(struct ublksrv_ctrl_dev *ctrl_dev, int evtfd, bool recover)
 {
-	struct ublksrv_ctrl_data cdata;
 	const struct ublksrv_ctrl_dev_info *dinfo =
 		ublksrv_ctrl_get_dev_info(ctrl_dev);
 	int dev_id = dinfo->dev_id;
@@ -330,7 +326,7 @@ static int ublksrv_device_handler(struct ublksrv_ctrl_dev *ctrl_dev, int evtfd, 
 	struct ublksrv_queue_info *info_array;
 	int i, ret = -EINVAL;
 
-	ublksrv_ctrl_data_init(ctrl_dev, &cdata, recover);
+	ublksrv_ctrl_data_init(ctrl_dev, recover);
 
 	snprintf(buf, 32, "%s-%d", "ublksrvd", dev_id);
 	openlog(buf, LOG_PID, LOG_USER);
@@ -361,7 +357,7 @@ static int ublksrv_device_handler(struct ublksrv_ctrl_dev *ctrl_dev, int evtfd, 
 	}
 
 	for (i = 0; i < dinfo->nr_hw_queues; i++)
-		sem_wait(&cdata.queue_sem);
+		sem_wait(&ctrl_dev->data->queue_sem);
 
 	ret = ublksrv_tgt_start_dev(ctrl_dev, dev, evtfd, recover);
 	if (ret) {
@@ -380,7 +376,7 @@ out:
 	if (ret)
 		ublksrv_ctrl_del_dev(ctrl_dev);
 	ublk_log("end ublksrv io daemon");
-	ublksrv_ctrl_data_exit(ctrl_dev, &cdata);
+	ublksrv_ctrl_data_exit(ctrl_dev);
 	closelog();
 
 	return ret;
@@ -592,6 +588,7 @@ static int ublksrv_cmd_dev_add(const struct ublksrv_tgt_type *tgt_type, int argc
 	//optind = 0;	/* so that tgt code can parse their arguments */
 	data.tgt_argc = argc;
 	data.tgt_argv = argv;
+
 	dev = ublksrv_ctrl_init(&data);
 	if (!dev) {
 		fprintf(stderr, "can't init dev %d\n", data.dev_id);
